# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a simple Nintendo Characters API",
    "version": "1.0",
    "title": "Nintendo Characters",
    "contact": {
      "email": "ghosty@blacksla.sh"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api",
  "schemes": [
    "http"
  ],
  "paths": {
    "/characters": {
      "get": {
        "tags": [
          "character"
        ],
        "summary": "Get all characters from the store",
        "description": "",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Character"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "character"
        ],
        "summary": "Add a new character to the store",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Character object that needs to be added to the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Character"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "character"
        ],
        "summary": "Remove a new character from the store",
        "description": "",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Character object that needs to be removed from the store",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Character"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid input"
          }
        }
      }
    }
  },
  "definitions": {
    "Character": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "game": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Character"
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}